create new  routes

+++ routes/api.php
@@ -11,6 +11,7 @@ use App\Http\Controllers\SearchController;
 use App\Http\Controllers\StatusController;
 use App\Http\Controllers\UserController;
 use App\Http\Controllers\SliderController;
+use \App\Http\Controllers\Admin\RoleController;
 use Illuminate\Support\Facades\Route;
 
 /*
@@ -29,6 +30,7 @@ Route::controller(UserController::class)->prefix('/user')->group(function () {
     Route::get('/{name}/{tag}', 'show');
     Route::get('/me', 'me');
     Route::post('/', 'store');
+    Route::post('/token', 'storeToken');
     Route::post('/update', 'update');
     Route::delete('/', 'destroy');
 });
@@ -40,17 +42,25 @@ Route::controller(AuthController::class)->prefix('/auth/token')->group(function 
     Route::delete('/{id}', 'destroy');
 });
 
-Route::controller(\App\Http\Controllers\Admin\UserController::class)->middleware('auth:sanctum')->prefix('/admin')->group(function () {
-    Route::post('/user/{id}', 'update_permission');
-    Route::post('/managerpermisions/{id}', 'manager_permisions');
+Route::controller(RoleController::class)->prefix('/admin')->name('admin.')->group(function () {
+    Route::post('/role', 'store')->name('role.store');
+    Route::get('/role', 'index')->name('role.index');
+    Route::get('/role/{id}', 'show')->name('role.show');
+    Route::put('/role/{id}', 'update')->name('role.update');
+    Route::delete('/role/{id}', 'destroy')->name('role.destroy');
+    Route::put('/role/permissions/{id}', 'updatePermissions')->name('role.permissions.update');
 });
 
-Route::controller(PeopleController::class)->prefix('/people')->group(function () {
-    Route::get('/', 'index');
-    Route::post('/', 'store');
-    Route::get('/{id}', 'show');
-    Route::delete('/{id}', 'destroy');
-    Route::post('/{id}', 'update');
+Route::controller(\App\Http\Controllers\Admin\UserController::class)->prefix('/admin')->name('admin.')->group(function () {
+    Route::post('/role/user/{id}', 'updateRoles')->name('role.user.update');
+});
+
+Route::controller(PeopleController::class)->prefix('/people')->name('people.')->group(function () {
+    Route::get('/', 'index')->name('index');
+    Route::post('/', 'store')->name('store');
+    Route::get('/{id}', 'show')->name('show');
+    Route::delete('/{id}', 'destroy')->name('destroy');
+    Route::post('/{id}', 'update')->name('update');
 });
 
 Route::apiResource('category', CategoryController::class);
@@ -59,24 +69,25 @@ Route::apiResource('status', StatusController::class);
 
 Route::apiResource('format', FormatController::class);
 
-Route::controller(MangaOverViewController::class)->prefix('/manga')->group(function () {
-    Route::get('/', 'index');
-    Route::get('/ids', 'indexIds');
-    Route::post('/', 'store');
-    Route::get('/{id}', 'show');
+Route::controller(MangaOverViewController::class)->prefix('/manga')->name('manga.over.view.')->group(function () {
+    Route::get('/', 'index')->name('index');
+    Route::get('/ids', 'indexIds')->name('index.ids');
+    Route::post('/', 'store')->name('store');
+    Route::get('/{id}', 'show')->name('show');
 });
 
-Route::controller(ScoreController::class)->prefix('/manga')->group(function () {
-    Route::post('/score', 'store');
+Route::controller(ScoreController::class)->prefix('/manga')->name('manga.')->group(function () {
+    Route::post('/score', 'store')->name('score.store');
 });
 
-Route::controller(MangaChapterController::class)->prefix('/chapter')->group(function () {
-    Route::get('/manga/{id}', 'index');
-    Route::get('/manga/all/{id}', 'indexAll');
-    Route::get('/ids', 'indexIds');
-    Route::get('/{id}', 'show');
-    Route::post('/', 'store');
-    Route::delete('/{id}', 'destroy');
+Route::controller(MangaChapterController::class)->prefix('/chapter')->name('manga.chapter.')->group(function () {
+    Route::get('/manga/{id}', 'index')->name('index');
+    Route::get('/lasts', 'indexLasts')->name('index.lasts');
+    Route::get('/manga/all/{id}', 'indexAll')->name('index.all');
+    Route::get('/ids', 'indexIds')->name('index.ids');
+    Route::get('/{id}', 'show')->name('show');
+    Route::post('/', 'store')->name('store');
+    Route::delete('/{id}', 'destroy')->name('destroy');
 });
 
 Route::controller(SearchController::class)->prefix('/search')->group(function () {
@@ -88,10 +99,10 @@ Route::controller(SearchController::class)->prefix('/search')->group(function ()
     Route::get('/people/{name}', 'showPeople');
 });
 
-Route::controller(SliderController::class)->prefix('/slider')->group(function () {
-    Route::get('/', 'index');
-    Route::get('/{id}', 'show');
-    Route::post('/', 'store');
-    Route::post('/{id}', 'update');
-    Route::delete('/{id}', 'destroy');
+Route::controller(SliderController::class)->prefix('/slider')->name('slider.')->group(function () {
+    Route::get('/', 'index')->name('index');
+    Route::get('/{id}', 'show')->name('show');
+    Route::post('/', 'store')->name('store');
+    Route::post('/{id}', 'update')->name('update');
+    Route::delete('/{id}', 'destroy')->name('destroy');
 });
